requirements.txt
python-docx
configparser
secure-smtplib
